services/reranker/requirements.txt
# Round-18 Task 9 cross-encoder reranker gRPC sidecar runtime deps.
grpcio==1.80.0
grpcio-tools==1.80.0
grpcio-health-checking==1.80.0
protobuf>=6.0.0,<8.0.0
sentence-transformers>=3.0.0,<4
torch>=2.2.0,<3
prometheus-client>=0.21,<1
